+++ src/lib/isAlphaChar.ts
