+++ src/lib/isArray.ts
