+++ src/lib/isBoolean.ts
