+++ src/lib/isTel.ts
@@ -1,7 +1,8 @@
 import ErrorMessage from '../utils/Error'
 
 /* eslint-disable no-useless-escape */
-const telRegex = /^\s*(?:\+?(\d{1,3}))?[-. (]*(\d{3})[-. )]*(\d{3})[-. ]*(\d{4})(?: *x(\d+))?\s*$/i
+const telRegex =
+  /^\s*(?:\+?(\d{1,3}))?[-. (]*(\d{3})[-. )]*(\d{3})[-. ]*(\d{4})(?: *x(\d+))?\s*$/i
 
 /**
  * Check if value is valid telephone number
